% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{clean_data_preparation_folders}
\alias{clean_data_preparation_folders}
\title{clean_data_preparation_tmp_folder}
\usage{
clean_data_preparation_folders(
  clean_tmp = TRUE,
  clean_local_storage = FALSE,
  clean_aws_storage = FALSE
)
}
\arguments{
\item{clean_tmp}{boolean Files and folders in local tmp will be deleted}

\item{clean_local_storage}{boolean Files and folders in local storage
will be deleted}

\item{clean_aws_storage}{boolean Files and folders in aws storage
will be deleted}
}
\description{
clean_data_preparation_tmp_folder
}
