% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_alignment_file}
\alias{prepare_alignment_file}
\title{Prepare alignment file for specified chromosome}
\usage{
prepare_alignment_file(
  alignment_id,
  chr,
  md5_data = NA,
  local_storage = FALSE,
  aws_storage = TRUE
)
}
\arguments{
\item{alignment_id}{char One of these possible values (100_way, 77_way)}

\item{chr}{chr alignment chromosome id}

\item{md5_data}{data.frame Data frame containing target md5 values
(default value NA)}

\item{local_storage}{boolean Save alignments to local storage
(default value FALSE)}

\item{aws_storage}{boolean Save alignments to aws storage
(default value TRUE)}
}
\description{
Prepare alignment file for specified chromosome
}
