% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_alignment_files}
\alias{prepare_alignment_files}
\title{prepare_alignment_files}
\usage{
prepare_alignment_files(
  alignment_id,
  chrs,
  local_storage = FALSE,
  aws_storage = TRUE
)
}
\arguments{
\item{alignment_id}{char One of these possible values (100_way, 77_way)}

\item{chrs}{chr Vector containing alignment chromosome ids}

\item{local_storage}{boolean Save alignments to local storage
(default value FALSE)}

\item{aws_storage}{boolean Save alignments to aws storage
(default value TRUE)}
}
\description{
Prepare alignment files for specified chromosomes
}
